pyproject.toml
[project]
name = "sakuragasaki46_freak"
authors = [
    { name = "Sakuragasaki46" }
]
dynamic = ["version"]
dependencies = [
    "Python-Dotenv>=1.0.0",
    "Flask",
    "Flask-RestX",
    "Python-Slugify",
    "SQLAlchemy>=2.0.0",
    "Flask-SQLAlchemy",
    "Flask-WTF",
    "Flask-Login",
    "Alembic",
    "Markdown>=3.0.0",
    "PsycoPG2-binary",
    "libsass",
    "setuptools>=78.1.0",
    "sakuragasaki46-suou>=0.2.3"
]
requires-python =  ">=3.10"
classifiers = [
    "Private :: X"
]


[tool.setuptools]
packages = ["freak"]

[tool.setuptools.dynamic]
version = { attr = "freak.__version__" }

